using the escaping for username
